move generation code
